ajouter que un utilisature qui n'pas inscrire ne peux changer un cours

+++ courses_list.php
@@ -1,9 +1,6 @@
  <?php 
  session_start();
- if(!isset($_SESSION['usere_id'])){
-     header('location:login.php?');
-             exit;
-}
+
  ?>  
 
    <?php include 'header.php'; ?>
@@ -63,13 +60,30 @@ EOD;
                     
                        echo  "<span><i class='fas fa-book-open'></i> $pos sections</span>
                         <span><i class='fas fa-clock'></i> Créé le $datecreet </span>
-                         <a href='enroll.php? course_id=$idcours ' class='btn-primary inscricours'><i class='fas fa-plus-circle'></i> S'inscrire</a>
+                          <a href='";
+                          if(isset($_SESSION['usere_id'])){
+                         echo"enroll.php";
+                         }else{
+                              echo "login.php";
+                         }  
+                       echo "'? course_id=$idcours ' class='btn-primary "; 
+                        if(isset($_SESSION['usere_id'])){
+                         echo"inscricours";
+                         }
+                       
+                       echo "'><i class='fas fa-plus-circle'></i> S'inscrire</a>";
+                         
+                 echo"
                     </div>
                         <div class='course-actions'>
-                        <a href='sections_by_course.php? course_id=$idcours ' class='btn-small'>Voir les sections</a>
-                        <a href='courses_edit.php? id= $idcours' class='btn-edit'><i class='fas fa-edit'></i></a>
-                        <a href='courses_delete.php? id= $idcours'  class='btn-delete'><i class='fas fa-trash'></i></a>
-                    </div>
+                        <a href='sections_by_course.php? course_id=$idcours ' class='btn-small'>Voir les sections</a>";
+                         if(isset($_SESSION['usere_id'])){
+                         echo"
+                         <a href='courses_edit.php? id= $idcours' class='btn-edit'><i class='fas fa-edit'></i></a>
+                         <a href='courses_delete.php? id= $idcours'  class='btn-delete'><i class='fas fa-trash'></i></a>
+                         ";
+                         }
+                echo"    </div>
                 </div>";
             }
                 ?>
@@ -80,4 +94,4 @@ EOD;
 
     <?php include 'footer.php'; ?>
 
-    <script src="script.js"> </script>
+
